+++ examples/modular-loading.js
@@ -1,0 +1,9 @@
+(async function () {
+  "use strict";
+  globalThis.modularEditor = await Editra.init({
+    selector: "#media-editor",
+    theme: "Word",
+    plugins: ["formatting", "table"],
+    toolbar: "foreColor highlighter | table undo redo",
+  });
+})();